requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
Werkzeug==2.3.6